requirements.txt
aiogram==2.19
python-dotenv==0.19.2
qrcode==7.3.1
pillow==9.0.1
asyncpg==0.25.0
